refactor(vertex-forager): split utils and unify exceptions
