+++ packages/vertex-forager/src/vertex_forager/exceptions.py
@@ -1,5 +1,21 @@
 from __future__ import annotations
 
+from dataclasses import dataclass
+
+try:
+    import httpx
+
+    HAS_HTTPX = True
+except ImportError:
+    HAS_HTTPX = False
+
+try:
+    import requests
+
+    HAS_REQUESTS = True
+except ImportError:
+    HAS_REQUESTS = False
+
 
 class VertexForagerError(Exception):
     """Base exception for all Vertex Forager errors.
@@ -133,6 +149,72 @@ class DataQualityError(VertexForagerError):
         super().__init__(f"Data quality validation failed for table '{table}' with rule '{rule}'")
 
 
+@dataclass(frozen=True)
+class RunError:
+    """Structured error information for pipeline runs."""
+
+    provider: str
+    dataset: str
+    symbol: str
+    exc_type: str
+    message: str
+    retryable: bool
+
+    @classmethod
+    def from_exception(
+        cls,
+        exc: Exception,
+        provider: str,
+        dataset: str,
+        symbol: str | None,
+    ) -> RunError:
+        exc_type = f"{exc.__class__.__module__}.{exc.__class__.__name__}"
+        message = str(exc)
+        retryable = cls._is_retryable_error(exc)
+        return cls(
+            provider=provider,
+            dataset=dataset,
+            symbol=symbol or "",
+            exc_type=exc_type,
+            message=message,
+            retryable=retryable,
+        )
+
+    @staticmethod
+    def _is_retryable_error(exc: Exception) -> bool:
+        if hasattr(exc, "response") and exc.response is not None and hasattr(exc.response, "status_code"):
+            status_code = exc.response.status_code
+            if status_code in {429, 503}:
+                return True
+            if 400 <= status_code < 500:
+                return False
+        if isinstance(exc, TimeoutError):
+            return True
+        if HAS_HTTPX and isinstance(
+            exc,
+            (
+                httpx.TransportError,
+                httpx.ConnectError,
+                httpx.ConnectTimeout,
+                httpx.ReadTimeout,
+                httpx.PoolTimeout,
+                httpx.NetworkError,
+            ),
+        ):
+            return True
+        if HAS_REQUESTS and isinstance(
+            exc,
+            (
+                requests.exceptions.ConnectionError,
+                requests.exceptions.Timeout,
+                requests.exceptions.ConnectTimeout,
+                requests.exceptions.ReadTimeout,
+            ),
+        ):
+            return True
+        return False
+
+
 __all__ = [
     "CheckpointNotFoundError",
     "ComputeError",
@@ -142,6 +224,7 @@ __all__ = [
     "InputError",
     "PrimaryKeyMissingError",
     "PrimaryKeyNullError",
+    "RunError",
     "SchemaMapError",
     "TransformError",
     "ValidationError",
